move code around once more to allow build with latest Go
Next release of Go will introduce rules for packages called "internal".
Also, this makes the import paths quite a bit shorter.
