+++ internal/dumpparser/main.go
@@ -1,10 +1,6 @@
-// Semanticizer, STandalone: parser for Wikipedia database dumps.
-//
-// Takes a Wikipedia database dump (or downloads one automatically) and
-// produces a model for use by the semanticizest program/web server.
-//
-// Run with --help for command-line usage.
-package internal
+// Guts of semanticizest-dumpparser, extracted into a module for end-to-end
+// testing.
+package dumpparser
 
 import (
 	"bufio"
